% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGroovy.R
\name{CheckJRERuntimeVersion}
\alias{CheckJRERuntimeVersion}
\title{Function verifies the existing JRE version is greater than or equal to 1.7 and, if it is not, invokes the stop
function with a message.}
\usage{
CheckJRERuntimeVersion(runtime_version)
}
\arguments{
\item{runtime_version}{For example, "1.7".}
}
\description{
Function verifies the existing JRE version is greater than or equal to 1.7 and, if it is not, invokes the stop
function with a message.
}
