% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGroovy.R
\name{Initialize}
\alias{Initialize}
\title{Function sets the global instance of GroovyShell that will be used by the Evaluate function whenever it is called
with a NULL GroovyShell parameter.}
\usage{
Initialize(binding = NULL)
}
\arguments{
\item{binding}{An instance of \href{http://docs.groovy-lang.org/latest/html/api/index.html?groovy/lang/Binding.html}{groovy.lang.Binding}.}
}
\description{
Function sets the global instance of GroovyShell that will be used by the Evaluate function whenever it is called
with a NULL GroovyShell parameter.
}
\examples{
 \dontrun{
 Initialize ()
 }

}
