% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{rGroovyAPI.env}
\alias{rGroovyAPI.env}
\title{An environment which is used by this package when managing package-scope
variables.}
\format{\preformatted{<environment: 0x000000002c53cc78> 
}}
\usage{
rGroovyAPI.env
}
\description{
An environment which is used by this package when managing package-scope
variables.
}
\keyword{datasets}

