% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Initialize}
\alias{Initialize}
\title{Function sets the global binding that will be passed to the GroovyShell
constructor.}
\usage{
Initialize(binding = NULL)
}
\arguments{
\item{binding}{The binding}
}
\description{
See also http://beta.groovy-lang.org/docs/latest/html/gapi/groovy/lang/Binding.html.
}

